Change Rational to typealias; rename generic type
This renames the generic type `Rational<IntegerBase>` to
`RationalNumber<IntegerBase>`, and adds a typealias
`Rational = RationalNumber<Int>`.

`ExpressibleByIntegerLiteral` cannot infer which type to use for
`IntegerBase` if it is not explicitly stated. For example:

    let x: RationalNumber = 88
    // error: cannot convert value of type 'Int'
    // to specified type 'Rational'

Renaming the generic type and adding a typealias allows the short name
`Rational` to be used in the simplest and most common case.